Fix for #134 : Armour penalties apply as bonuses
Armour penalty to speed was always added to speed, armour penalty to skills was always subtracted from skills.  This make for inconsistencies where a -1 speed reduced speed, but -1 skills boosted skills.

Have fixed this by always taking Math.abs() and then subtracting it, so it doesn't matter whether the penalty emptied for move or skills is -ve or +ve

Have tested with multiple types of armour, together and with shields.

Also minor update to Grab and drag powers to bring the text more in line with the powercard text.