showing terms
This is is a bit of a hack solution, but it fixes a messy interaction between how class MultiAttackRoll Class relates back to the Roll Class, where how it is set up an extra hidden roll is taking place and that is being placed MultiAttackRoll.terms. The Dice So Nice module then only picks up this extra roll, and displays that with the 3d dice, so users are left with only this singular unrelated dice roll instead of all the other dice they were expecting to have rolled. 
This also fixes part of an issue with chat.js highlightCriticalSuccessFailure, which again would only look at this extra roll when determining to highlight for critical success or failure. So the entire multi roll log would be highlighted red or green at random internals.