Consumables with Attack & Damage Rolls
Gave consumables a cutdown version of a powers attack roll.  Chunk of copy and paste for the UI and refactored the system.json template to have attack and Damage data as a common template that can be shared.  That meant that all the power-card generation stuff /just worked/ (with a very few tweaks)

Updated the existing potions to the new template, and added poisons and a smattering of alchemical items to test.  Didn't want to add all of them until we are happy with the new implementation.