Use effects for Custom @variables in attack and damage formula
Can now specify an effect that is @variable and refer to that effect in attack and damage formula nad it will be substituted in.

Usage: Effect key must be "@varname" the code triggers on the effect key starting with @  Multiple effects with the same key will add together, but that is not really supported operation.