Compendium Equipment Update:
SRD Implements (Staff, Wand, Rod) updated to be tagged with the relevant implement group.  Also removed weapon proficient from wand and rod as they are not weapons (but staffs keep it, and even get a little flavour text)

Also removed the SRD Shared powers compendium from system.json, as it was empty with all powers moving to example.