Update to effect variables
custom variables and attack/damage effects can now use @variables - the standard set that are used in attacks / damages.

By quirk this also allows for attack/damage effects to use custom variables.

attack/damage effects have to evaluate a non numerical expression to compute if its larger or not, but custom variables can just use the appended string method that @EndlessNights factored in.

Moved the custom variable substitution to last and did a pre check to see if there were any @variables left to improve performance - there is no need to iterate the effects on a character if there are no variables left to substitute.