Figured out why Item macros kept getting massive numbers of indents.
It was picking up the indentation of the link used to insert the Macro Tab into the html.

The first line didn't pick it up, because the first line was immiedately following the opening <TextArea> tag.  However, I am guessing because of some interaction between how things inserted by the templater behave, all subsequent lines, which are subsequent lines in the html (as textarea uses its value as the contents) picked up the indentation of the template invoker.

Have de-indentented and left a comment on all the places that reference the macro template.