Added Hooks for Token Action HUD
When adding equipment to token action HUD integration I was reminded that in order to get it working I had to poke some quite internal and knarly methods in 4E.

Basically things that we would change, that would then break token hud, and we would have no idea why.

So I refactored them all into the hooks.js object, which gives Token HUD a single place for when it needs to call any method on one of our objects.  Especially as a lot of them the logic is actually in the sheet.  This means they will at least show up in a usages search in VSCode/Webstorm etc... so we know if we change them to also update the calling method.

Method signatures were made to be as standardised as possible with all relevant information, so hopefully if we ever change an implementation there is no need to change api.