Steve gets an upgrade
Steve has gained an implement and an AoE fire power that uses the implement.

An example ritual and an example NPC (Badgers are not in 4E as creatures, so do not have to worry about OGL issues) (Ritual costs are tricky to grab as the attribute dropdown gets cleared as soon as it's off a character)

Feat that adds to fire attack and damage powers (to show on steves new fire power).

Updated fighter weapon talents to have an effect to boost that attack. This can be done automatically for 2 Handed Weapons, alas 1 handed weapons need a custom variable, but at least we get a custom variable demo!