Added Level to all inventory items as a top level field, like on paragon/epic features
also allows @itemLevel to reference it.

updated example potions and example alchemy to have the correct values in that field (they are the only example things I know of that have a level, everything else is non magical).

Really wanted to add @enhance to equipment, but realised that was a can of worms that was not quick to implement.