Update manual.db
Going over the manual and adding additional details and a few other minor tweaks.

In general to any image, I added an additional 2px to the height and width. The way Journals render images with their boarders, they add a boarder which causes the image to be scaled inwards slightly. Which in turn causes some of the final detailing to become a little blurry. A minor thing, but just lets the small text retain some of its sharpness. As well as some minor spelling mistakes. Also didn't touch any of the British v American spell.

Attacks and Rolls # Rolling Multiple D20's
Appended to include the following details
> User may be able to quickly swap between setting up the dice to rolling two dice by holding down CTRL + Left Clicking either the plus or minus buttons which will immoderately set the value to either -2 or +2
> 
> With a value of zero begin set for the number of D20's to be rolled, no D20's will be rolled, and the results will derive from whatever other bonus where associated with the roll.

Character Sheet # Hit Points
Appended to last line to include a reference to the Tough Feat
> Along with further modification from the @Compendium[dnd4e.example_feats.QifwVUaX7Mkq54eL]{Tough} feat which grants an additional 5 hits per character tier level.

Character Sheet # Speed and Initiative
Appended to include:
>  Mouse-Hovering over the either the "Speed value" or "Speed button" will reveal a tooltip that indicates all of a characters calculated speed values for common movement actions they can take.
> ...
> Clicking on the Initiative Value to bring up a dialog window which will let users be able to adjust bonus to your initiative roll, as well as which Ability Mod is used in association with the roll. By default, this is set to Dexterity.
> 
> By clicking on "Initiative" while their is currently an active encounter. You will roll initiative and your character will be added to the Combat Tracker. You may also notice that a decimal value is added to your Initiative Roll, this is equal to 1/100th of your Initiative bonus, and is used as a quick way to resolve ties in the Initiative Roll results.

Character Sheet # Skills
Appended to include the following details
> At the bottom of the skill list, you will see a button labeled "Edit Skill & Ability Custom Descriptions." Clicking this will open up a window where you may edit what the default flavor text is for your Ability and Skill rolls. The default text is "@name uses @label." Where @name is replaced by the name of your character, and @label is replaced by the name of the corresponding Ability or Skill.