Formula term highlight rollbac + minor fixes.
attempts to fix Foundry's occasional editing of formula made things look gash with variables in the output, so I've reverted that.

It works most of the time, but occasionally Foundry will change the formula by adding or removing a space.  Note for Me for future: It was Rise, Criticals with the greataxe.  The axe critical bonus formula "(@enhance)d6 + @wepDice(@tier)"  for some reason when foundry parses it, it spits back out "(@enhance)d6 +@wepDice(@tier)" (though with the numbers obviously)

Given its much more edgecasey I am leaving it (and all that happens is the highlighting won't get individual terms).   have 2 solutions in the pipe: 1. let the highlighter fail in a more granular way, 2 enforce spacing around operators, but that could be perilous.

 Other changes:
 Moved the effect bonuses for damage to be part of the main expression so they get the damage type of the primary damage type, as they were breaking resistance calculations otherwise.  Note this is still a problem (and always has been) with bonus damage.  I think there is a bug or 2 in the damage type / resistances calculator, but that is for another time.

 Minor change to replacer to handle nulls:  If you had selected "None" as weapon but damage type as "base weapon damage" and still had @wepDamage in the formla (a possible combination given defaults (invalid, but possible)) the helper with crash with an NPE.  This way you get an output (that looks awful, so you go and fix it)