#176 : Action Point Riders
A chunk of this was already done (guessing as 5e holdover?), just needed a dialog to enter the extra riders and the code to output them was pre-written.

No automation on this as there are so many different types of AP benny you can get from classes (warlords) and Paragon Paths, that trying to integrate them all would be a mugs game.