Minor Fixes:
Consumable name and type are part of main item Document not system collection, fixes "undefined undefined" when using consumables

Effects would not set for the first combatent because combat.turn = 0 === false when put in a truthy comparison check.  I actually managed to find this in a bizarre edgecase where the next line failed because combat.turn had been set, but there was no combatent in that part of the array so got an undefined.  Have not been able to replicate it, but made the check more paranoid.

Heal dialog menu now displays surge value.  (mostly useful for me because my players need to deal with mournland healing which halves the value)

changelog updated to fix spelling and fix mis-attributed changeset.