Fix for #248
Rejigged the _getItemConsumptionTargets method:

'attribute' is now first, because attributes can now function separately.
If there is an attached actor, function as before.
If there is no actor grab the Player Character model and use that.

If you've not asked for an attribute and are not attached to an actor, then return nothing as before, the others all need a backing actor.

Items with attribute backing can now be created in the items tray and copied backa nf forth between characters and items so long as they are using default model.