Updated various headings and labels
I noticed that the headings for things like attack and damage formulae didn't differentiate between weapon and implement. I think they used to? So I probably wrecked with my big translatable strings update. Anyway, now each appropriate label specifies "weapon" or "implement", but only if the weapon is BOTH a weapon and an implement. Otherwise it omits that part of the string.