Updates to Manual:
-Made use of pages
-A lot of different manuals compacted down into the character sheet manual page
-Fixed the links

Compendium checkins:

Advice on discord: commit entire directory. Eurgh :(

Been observing and reading around level DB:  https://www.cclsolutionsgroup.com/post/hang-on-thats-not-sqlite-chrome-electron-and-leveldb gives a good explanation (skip down to LevelDB Overview)

So probably don't need the latest up to date LOG file.  Probably, but no harm in it.

We don't need to commit *.log files, provided that when you edit a compendium, you close and restart foundry once afterward, which seems to convert the latest .log file into an ldb file regardless of size.  I think, based on what I observed.

We will need to commit .log files that contain any data if they don't get cycled into ldb files.  For this reason I've taken them off the .gitignore, because you want to be able to see them when checking commits to see if they are empty (ignorable) or with data (should be commited)

At some arbitary point levelDB will decide to compact the ldb files down to 1.  So someone is going to have a bit of a high stress commit.

It also means that every time one of us actually changes a db, and thus needs to commit the manifest file, everyone is going to have a nasty merge :(