Removed "All" and "Ongoing" from selectable
In the checkbox section for primary damage type, and the drop-down for secondary damage type and secondary crit damage type, prevents the All/Damage type and the Ongoing type from being listed. The reason of course being that these types are used only for resistances, and no attack should actually need them to model its damage correctly (ongoing damage being an invisible type in effects, and not handled in the power details tab).