Flavour/desc variance for NPCs
Wigmeister's fix reminded me of an issue where NPC power cards can output unwanted text when they are set to auto. Specifically, they follow the same logic as player powers, and output the full description if there is no "chat flavour". This makes sense for PCs, since almost all of their powers have flavour, but it's not ideal for NPCs because their powers almost never have flavour text.

This update changes the logic so for /auto-generated cards on NPCs only/, the flavour section will be skipped entirely if "chat flavour" is empty (and simplifies the template code a fair bit, as a bonus). Non-auto cards are unchanged.