Apply Effects Buttons
For when the power has hit/miss/both effects, but the player didn't target the enemies so they don't get applied (or the wrong one gets applied etc...)

Appear below the "attack"/"damage" buttons, the UX isn't great, but its functional.

Because for the GM, or at least for me, it's easier to do this with selections than with targets, I made a system option that flips the targetting for these buttons from targeted tokens (default) to selected tokens.