Refactors for TAH integration.
Moved the groupings from inline objects in the actor-sheet to constants in config to let me reference them in the TAH integration.

moved the work of actually generating the html summary of an item out of _onItemSummary and into its own method _generateItemSummary.

TAH hooks.js updated with some new methods to make use of the exposed method and keep power sorting as close to the sheet as possible.