Restored res.armour property, emptied sustain action
I was a bit overzealous in removing the "armour" property from the res properties and breaking new actors, my bad.
While I was here, I removed "none" as the detail power sustain action, since that maps to No Action rather than no sustain. The default is just empty, which is appropriate for the majority of powers.