Fix: right click Apply Damage/Healing not working
Ironically the check was causing the issue as it was not coping with the fact that the Terms of a roll have mulitple different types, and the roll property varies between them.

while roll? was fine for a flat numeric term, it was an async method on a Die term.  Safer just to use the .total property, which always returns the evaluated total of the term, and then check if that is numeric or a string.

Similarly we were digging into the options to get the flavour of a roll, which is undefined if there was no flavour, leading to NPE's in the healing surge calculation if you tried to heal without using a flavoured roll.  Changed to the API and added a null check, though  think the API always seems to return empty string, but there is no guarentee of that on the docs so put a ? in there.