Fix: Token Size not being updated
token property on actors moved to prototypeToken, so player actors and important monsters will now scale again nicely.

Note that this will not resize existing tokens for a linked actor when you change it.  While we could go down there using the getDependantTokens property I decided that was too deep down a rabbit hole for the update method on the actor.