Updated ongoing damage handling to allow dice
Improved handling of ongoing to damage to allow dice (and mixed dice/number/variable) values.
As currently designed, rollable values /do/ need to be evaluated before the chat report happens, as we compare DoTs first in order to to choose the highest value; that means the roll happens when the chat card is sent, without any user input. This behaviour will probably be surprising when the automation damage is set to manual/reminder, but I don't think there's a good way around it right now, so I'm calling it okay?