Change resource value to text field
Resource depletion worked correctly with a string instead of a number, but the resource would still break if the character sheet was closed (or VTT reloaded, or possible other reload/update conditions) with a 0 value. I suspected this was because it would still read the field and convert it to an int on close (then interpret 0 as false on load?), so I swapped the sheet field to a text input as well. This seems to have done the trick!