Updates resources default state
In case it was interfering with the resource consumption bug, I've changed the default state for resources to be empty strings for value/max and `false` for the two checkboxes. Might not be strictly necessary, but it doesn't seem to have hurt anything either, so not a bad idea I think?