Made "armourBaseTypeCustom" display for all armour
Pointed out by Cole Grippah on Discord—previously this field was dependent on selecting "custom" for base armour, but this option seems to be obsolete, so it was never displayed except for on legacy items. Meanwhile, the field has value for specifying a "special" base armour type on top of the normal base, such as for superior or masterwork armours. It seems logical to just let it display for all armour types.