Updated consumable behaviour
If a consumable doesn't delete on empty AND has quantity 1 AND has limited usage based on Encounter/Day/Round, it's probably meant to be a singular item which replenishes charges, rather than a stacking consumable that should be depleted. However, the current behaviour depletes the quantity for such a item, meaning uses will not recharge on the specific period. This update makes it so the quantity will never be depleted IF all those conditions are met.