Adjusted text regen chat card
Chat card produced by regeneration effect will no longer say "ongoing damage" but "regeneration" instead. Additionally, no longer print a chat card if regen value is 0; there's value in seeing that ongoing damage was resisted, but I feel it's just clutter when it's regen. Example circumstance under which it would be 0 is an effect that only grants regen while bloodied; value gets multiplied by `@bloodied` which is 0 when not bloodied. I don't think anyone wants a chat card every turn in that situation. But it is definitely a more subjective change, which is why I'm explaining myself, haha