Add field to Powers for "Attack Text Details"
This field is to accommodate powers such as the rogue's Jumping Blade Assault and Shadow Steel Roll, which include additional details in their Attack lines. The existing behavior is defaulted to if nothing is entered into this field.

'@attackValues' and '@attackTotal' are provided as helpers so that the existing tooltip behavior is still possible with custom text entered into this field.