Add "Save Against Effect" dropdown to save dialog
Adds a dropdown to the save dialog when rolled from the character sheet allowing the user to select an effect to save against. This dropdown only appears if there are effects that can be ended on a save and the save wasn't already being prompted for a specific effect (as with auto saves).

Currently the DC doesn't update when you select an effect. I'm trying to figure out how to make that happen.