Add damage types from ongoing damage entries
Typically these will already be present in `suitableKeywords` due to them being, well, keywords from the power that created the effect. But there's no harm in guarding against a user forgetting to add a damage keyword at the power level. This also offered an opportunity to explicitly add an "untyped" internal (i.e. entirely local to the bonus processing function) keyword just as a matter of convenience for users who are trying to implement features that grant bonuses to saves against ongoing untyped damage (save them from having to make the mental leap from untyped to physical).

Additionally, autoDOTs would fail ungracefully if the user forgot to explicitly add a damage type to it, so in such a case that DOT is now treated as physical/untyped damage, similar to normal power damage rolls.