Movement improvements
Burrow, fly, and teleport are added as movement types.
Run is reconceptualized as not a speed in and of itself but as a bonus, since in the rules it can be combined with any speed you happen to be using when you take the Run action. Implementation-wise nothing changes with it except that the default is now simply 2.
TokenRuler4e takes into account a token's current movement mode and run bonus to determine what color to use for paths.
Special speeds like climb and swim now default to 0; TokenRuler4e implements the additional cost for swimming or climbing without a swim or climb speed. The same behavior extends to crawling, which isn't a separate speed but exists in the rules and in Foundry core.